+++ src/framework/GoogleTestSuiteInfo.ts
@@ -228,7 +228,12 @@ export class GoogleTestSuiteInfo extends AbstractTestSuiteInfo {
               "Couldn't parse output file. Possibly it is an older version of Google Test framework. It is trying to parse the output",
             );
 
-            this._reloadFromStdOut(googleTestListOutput.stdout, oldChildren);
+            try {
+              this._reloadFromStdOut(googleTestListOutput.stdout, oldChildren);
+            } catch (e) {
+              this._shared.log.info('GoogleTest._reloadFromStdOut error', e, googleTestListOutput);
+              throw e;
+            }
           }
         }
       });
@@ -355,10 +360,10 @@ export class GoogleTestSuiteInfo extends AbstractTestSuiteInfo {
         }
       };
 
-      runInfo.process!.stdout!.on('data', (chunk: Uint8Array) => processChunk(chunk.toLocaleString()));
-      runInfo.process!.stderr!.on('data', (chunk: Uint8Array) => processChunk(chunk.toLocaleString()));
+      runInfo.process.stdout!.on('data', (chunk: Uint8Array) => processChunk(chunk.toLocaleString()));
+      runInfo.process.stderr!.on('data', (chunk: Uint8Array) => processChunk(chunk.toLocaleString()));
 
-      runInfo.process!.once('close', (code: number | null, signal: string | null) => {
+      runInfo.process.once('close', (code: number | null, signal: string | null) => {
         if (code !== null && code !== undefined) resolve(ProcessResult.createFromErrorCode(code));
         else if (signal !== null && signal !== undefined) resolve(ProcessResult.createFromSignal(signal));
         else resolve({ error: new Error('unknown sfngvdlfkxdvgn') });