+++ test/RealExecutables.test.ts
@@ -130,7 +130,23 @@ describe(path.basename(__filename), function() {
 
       const eventCount = adapter.testStatesEvents.length;
       await adapter.run([adapter.root.id]);
-      assert.strictEqual(adapter.testStatesEvents.length - eventCount, 92, inspect(adapter.testStatesEvents));
+
+      const labels: string[] = [];
+      adapter.testStatesEvents.forEach(x => {
+        if (x.type == 'test' && x.state == 'running' && typeof x.test != 'string') labels.push(x.test.label);
+      });
+
+      assert.strictEqual(
+        adapter.testStatesEvents.length - eventCount,
+        2 + // start,fin
+        2 + //suite1 start,fin
+        5 * 2 + // suite1 tests
+        2 + //suite2 start,fin
+        2 * 2 + // suite2 tests
+        2 + //suite3 start,fin
+          35 * 2, // suite3 tests
+        inspect(labels),
+      );
     });
 
     it.skip('should be notified by watcher', async function() {